Use uint32_t because that's what the class uses internally.

+++ src/mlpack/core/math/random.hpp
@@ -32,12 +32,14 @@ namespace math /** Miscellaneous math routines. */ {
 
 /**
  * Set the random seed used by the random functions (Random() and RandInt()).
+ * The seed is casted to a 32-bit integer before being given to the random
+ * number generator, but a size_t is taken as a parameter for API consistency.
  *
  * @param seed Seed for the random number generator.
  */
 inline void RandomSeed(const size_t seed)
 {
-  randGen.seed(seed);
+  randGen.seed((uint32_t) seed);
 }
 
 /**
